Baseline and release are separate operations (#207)
Baselining allows us to compare a pull request to master based on code coverage and security, so we can see what changed. However, these shouldn't block a release, because we have already tested it. Also, pull request merges and commits to master are conceptually different in git, so different workflows give a simpler model.