Bugfix: Chinese definition contains word language (#260)
## Problem
When looking up definition results from elasticsearch, we need the definition language, the word language, and the definition source. However, ChineseDefinitions don't take the word language as a parameter because it is obvious. This means that it is not serialized to JSON, and is missing from elasticsearch.

## Solution
Manually add definition language to the fields written to JSON

## Also
Webster domain types are kept in a duplicate package: `webster.webster`. I moved them to the correct spot.