Feature: Store definition entries instead of definitions in cache (#262)
## Enhancement
We store our processed definitions for each source in elasticsearch. However, if we ever add new fields, then we'll have to reprocess all the definitions. Instead, let's store the definitions raw and pull the relevant fields out at load time. 

### Won't that be slow?
We can add a cache later on if performance becomes an issue.

### New pattern: Fetchers
Fetchers hold all the logic for getting definitions from a specific source. That way definition services don't need to know the how of getting definitions, just which ones to call.