Wrap Elasticsearch caching in a container object (#263)
## Problem
There have been multiple bugs where an object to be cached doesn't write out the field needed to find itself again. This burden is passed on to the programmer. 

## Solution
Wrap the items in a class which contains the search field. This guarantees it is always saved in elasticsearch, and thus findable.