Don't rebuild dependencies image if nothing has changed (#274)
## Problem
Every version changes the build number in `build.sbt`. This means the dependencies image is always different, and needs to be rebuilt.

## Solution
Provide a dummy `build.sbt` file which does not contain the version, and doesn't change. Use this for building dependencies.

## Why?
Cuts a 3-5 minute step from every build, speeding up PR checks and deployments.