Upgrade all Alpine 3.12 packages to the latest versions (#449)
The openjdk version we use `openjdk:15-ea-jdk-alpine3.12` is no longer maintained, with the earliest supported version being JDK 17. However, guice doesn't support Java 17, and thus we are blocked.

In the long term, the solution is to wait for Guice to fix the issue (identified in September), or find dependencies that are more up to date.

In the short term, we can mitigate these problems by just manually upgrading insecure packages. This closes 60+ vulnerabilities, so it's worth doing. This doesn't fix the already broken build, but it brings us a lot closer.